requirements.txt
discord.py==1.3.4
python-dotenv==0.14.0
black
discord.py[voice]
pytube3==9.6.4
youtube-search==1.1.0